Add more GetMovementDirection tests and fix bug
I added the MovementWithZeroButtonsPressed, MovementWithTwoButtonsPressed, and MovementWithFourButtonsPressed tests. In the process of adding these tests, I fixed a bug where the player wouldn't move correctly if the player was hitting an obstacle. The bug fix allows the player to "slide" around obstacles if more than one movement button is pressed.