Implement attack animation
Plus, fix usage of Scale. The game will be scaled up (currently at 3x) from the original game. This is because I want to use integer values for coordinates and speeds, but at the original scale (16x16 pixel sprites), there's not enough pixels on the screen to have an integer WalkSpeed and not have the game look ridiculous, i.e. at Scale = 1, WalkSpeed = 1 is too slow, and WalkSpeed = 2 is too fast.