Implement collision with immovable objects
In order to make implementing collision with immovable objects easier, I added two abstractions: Sprite and World. Sprite is just a convenient way to store the information needed to draw a sprite on the screen. The World is an abstraction for everything in the game world: The player, the immovable objects, the enemies, the projectiles, etc. Basically anything that Link would see is included in the World. This does not include the HUD, however, since Link would not be able to see the HUD. The HUD is drawn on top of the World.

Also, Visual Studio was bugging me about adding the Microsoft.CodeAnalysis.FxCopAnalyzers analyzers, so I added them.