Narrow the player's walking collider
This matches the original game more, and it makes it *much* easier to walk through openings that are only one tile big.

Also in this commit: Add a rock boundary around the demo room, with one opening. This opening will be used in a later commit to test warping between rooms.