Update MonoGame to v3.8.0
And update other libraries (test libraries, MonoGame Extended) as well.

This commit deletes the Content\dll directory in order to update the MonoGame Extended library. Adding a reference to Content.mgcb to a directory outside of the project is not great though, so maybe in a future commit I will fix this. The root issue, https://github.com/craftworkgames/MonoGame.Extended/issues/626, is still not fixed.