Add "slide around corners when walking" feature
Being able to easily walk through tight corridors is absolutely necessary for this game. Otherwise, it will be hard and frustrating to walk between rooms in a dungeon. The NES Legend of Zelda addresses this problem by having a "rails" system, where the player can only walk on invisible rails in the map. The Legend of Zelda: A Link to the Past (ALTTP) addresses this problem differently by allowing Link to slide around corners when walking. I think the ALTTP way is the better, more modern way to address this problem, so that is what this commit does--add the "slide around corners when walking" feature.